requirements.txt
# 13 APR 2018 UPDATED

# Need to install below apt-get packages before
# sudo apt-get install python-pip libxml2 libxml2-dev libxslt1-dev libmysqlclient-dev uwsgi nginx
# sudo apt-get build-dep python-imaging
# follow 2 steps link https://docs.confluent.io/current/installation/installing_cp.html#installation-apt
# sudo apt-get install librdkafka++1 librdkafka-dev librdkafka1 confluent-kafka-2.11

amqp
#anyjson
attrs
Automat
autopep8
billiard
# blist
boto
cassandra-driver
# cassandra-snapshotter==1.0.0
celery
chardet
confluent-kafka
constantly
crcmod
datetime-truncate
decorator
defusedxml
dse-driver
Django
django-allauth
django-bootstrap-form
django-bootstrap3
django-braces
django-cassandra-engine
django-compat
django-contrib-comments
django-cors-headers
django-crontab
django-debug-toolbar
django-easy-pdf3
django-extensions
django-guardian
django-hijack
django-hijack-admin
django-markdown-deux
django-organizations
django-pandas
django-recaptcha
django-registration-redux
django-rest-auth
django-rest-framework-docs
django-rest-swagger
django-screamshot
django-silk
django-tagging
django-timedeltafield
djangorestframework
djangotoolbox
django-urls
docutils
Dozer
drf-nested-routers
ecdsa
enum34
et-xmlfile
Fabric
funcsigs
futures
geographiclib
geopy
google-compute-engine
gprof2dot
# gyp
html5lib
httplib2
hyperlink
idna
incremental
ipython
ipython-genutils
isoweek
jdcal
Jinja2
jsonfield
jsonpickle
kombu
# librabbitmq
lockfile
lxml
Mako
markdown2
MarkupSafe
mock
# MySQL-python
mysqlclient
nose
numpy
oauthlib
openpyxl
pandas
paramiko
path.py
pathlib2
pbr
pexpect
pickleshare
Pillow
postmarker
psutil
ptyprocess
pycodestyle
pycrypto
Pygments
pymodbus
PyPDF2
pyserial
python-daemon
python-dateutil
python-magic
python-memcached
python-openid
pytz
PyYAML
redis
reportlab
requests
requests-oauthlib
roman
s3cmd 
scandir
scikit-learn
scipy
simplegeneric
simplejson
six
SOAPpy
Sphinx
sqlparse
tabulate
traitlets
twilio
Twisted
# Unirest
urllib3
uwsgitop
virtualenv
WebOb
wstools
xhtml2pdf
xlrd
XlsxWriter
xlwt
zope.interface